fix: Wait for replication write on all replicas
Tries to wait for the replication message to be written on all
replicas before returning, that way we make sure data is on
all replicas on a healthy cluster.

Additionally, it uses the page cache for the max offset file,
which is suitable for this case.

It contains lint fixes for error checks and some test fixes.